+++ src/ngraph/frontend/onnx_import/ops_bridge.cpp
@@ -35,6 +35,7 @@
 #include "op/ceil.hpp"
 #include "op/clip.hpp"
 #include "op/com.microsoft/dequantize_linear.hpp"
+#include "op/com.microsoft/quant_conv.hpp"
 #include "op/com.microsoft/quantize_linear.hpp"
 #include "op/concat.hpp"
 #include "op/constant.hpp"
@@ -72,7 +73,6 @@
 #include "op/pad.hpp"
 #include "op/pow.hpp"
 #include "op/prelu.hpp"
-#include "op/quant_conv.hpp"
 #include "op/reciprocal.hpp"
 #include "op/reduce.hpp"
 #include "op/relu.hpp"
